requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
Pillow>=10.2.0
ddddocr>=1.4.11
pytesseract>=0.3.10
